Error Rapot

+++ prestasi-app/database/migrations/2024_05_24_144840_create_rapots_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
             $table->foreign('id_wali_kelas')->references('id')->on('wali_kelass')->onDelete('cascade')->onUpdate('cascade');
             $table->bigInteger('id_mapel')->unsigned();
             $table->foreign('id_mapel')->references('id')->on('mapels')->onDelete('cascade')->onUpdate('cascade');
-            $table->float('nilai_pengetahuan', 5)->nullable();
+            $table->integer('nilai_pengetahuan')->nullable();
             $table->char('huruf_pengetahuan', 1)->nullable();
-            $table->float('nilai_keterampilan', 5)->nullable();
+            $table->integer('nilai_keterampilan')->nullable();
             $table->char('huruf_keterampilan', 1)->nullable();
             $table->timestamps();
         });
